+++ src/datarobot_genai/drmcp/core/config.py
@@ -145,7 +145,7 @@ class MCPServerConfig(DataRobotAppFrameworkBaseSettings):
         description="YAML configuration for OAuth protected resource metadata",
     )
     mcp_enable_unauthenticated_well_known_route: bool = Field(
-        default=False,
+        default=True,
         description="Enable/disable unauthenticated well known route",
     )
     # When the server is run in a custom model, it is important to mount all routes under the